Reverted to by-subcorpus analysis
Why did I ever decide NOT to do this?
The corpora are smaller and not perfect, but by using all the data, which were designed for other purposes, there is far too much skewing of the analysis. For example, 55 examples of ana_2 and ana_3, but hundreds for ana_1 and ana_0...